Improves report template completeness analysis / Block complex analysis on partially template.
Adds a check to limit the groups displayed in the completeness analysis plot to a maximum of 6, preventing errors when too many groups are present.

Check the complexity of the model and thrown an error when it is more than a pairwise comparisons. Support for formula  '~ group' is allowed .

Removes redundant ggplot functions.

Adds logging for absent analysis.